moved images to carousel page to make the component more resueable
